requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# HWP Processing
pyhwp==0.1b15
# Alternative: hwp5==0.1.0

# PDF Generation
reportlab==4.2.2
# Alternative: weasyprint==62.3

# Database & ORM
sqlalchemy==2.0.31
alembic==1.13.2
asyncpg==0.29.0  # For PostgreSQL async support
psycopg2-binary==2.9.9  # PostgreSQL adapter for Python

# Task Queue
celery==5.4.0
redis==5.0.7
flower==2.0.1  # Celery monitoring

# Storage (Phase 3)
# boto3==1.34.144  # For S3 storage

# Validation & Serialization
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0  # Email validation for pydantic
python-magic==0.4.27  # File type detection

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
slowapi==0.1.9

# Logging & Monitoring
structlog==24.4.0
prometheus-client==0.20.0
psutil==7.0.0

# Development & Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
httpx==0.27.0  # For testing async endpoints
black==24.8.0
ruff==0.5.6
mypy==1.11.1

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0  # Async file operations
olefile==0.47  # For HWP file parsing
fpdf2==2.7.9  # Better Unicode/Korean font support
PyMuPDF==1.24.10  # For PDF manipulation and fixing
pdfplumber==0.11.4  # For PDF text and table extraction